+++ app/src/main/java/com/sismantec/conteoinventario/modelos/DetalleConteoJSON.kt
@@ -2,6 +2,6 @@ package com.sismantec.conteoinventario.modelos
 
 data class DetalleConteoJSON (
     val id_Inventario: Int,
-    val existencias: Float,
-    val existencias_u: Float
+    val existencias: Int,
+    val existencias_u: Int
         )
